Remove a bunch of cruft from %Frame{}
