+++ lib/core/utils/utils.ex
@@ -1,5 +1,4 @@
 defmodule ScenicWidgets.Utils do
-
   @doc """
   Checks if a coordinate point (a tuple in the form `{x, y}`)
   is inside a bounding box, defined in the same format as
@@ -9,6 +8,6 @@ defmodule ScenicWidgets.Utils do
     # NOTE: Because the y axis starts at zero in the top-left, and
     # gets larger as we go down the page, it's a little counter-intuitive
     # to calculate if we're inside the bounds
-    (x >= left and x <= right) and (y >= top and y <= bottom)
+    x >= left and x <= right and (y >= top and y <= bottom)
   end
 end
